requirements.txt
youtube_transcript_api
streamlit
google.genai
python-dotenv
pathlib